genai-requirements-automation_gradio/requirements.txt
fastapi
uvicorn
gradio
openai
PyPDF2
python-multipart
requests
pydantic
jinja2
python-dotenv
reportlab
python-docx